Changed to Classlib Project
