--- ClassesAutoGen/Plantel.cs
@@ -1,43 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-using Microsoft.EntityFrameworkCore;
-
-namespace Proyecto_Almacen.AutoGen;
-
-[Table("Plantel")]
-public partial class Plantel
-{
-    [Key]
-    [Column(TypeName = "int")]
-    public int? PlantelId { get; set; }
-
-    [Required]
-    [Column(TypeName = "nvarchar(50)")]
-    [StringLength(50)]
-    public string Nombre { get; set; } = null!;
-
-    [Required]
-    [Column(TypeName = "nvarchar(100)")]
-    [StringLength(100)]
-    public string Direccion { get; set; } = null!;
-
-    [Column(TypeName = "int")]
-    public int? Telefono { get; set; }
-
-    [InverseProperty("Plantel")]
-    public virtual ICollection<Almacenista>? Almacenistas { get; set; }
-
-    [InverseProperty("Plantel")]
-    public virtual ICollection<Coordinador>? Coordinadores { get; set; }
-
-    [InverseProperty("Plantel")]
-    public virtual ICollection<Docente>? Docentes { get; set; }
-
-    [InverseProperty("Plantel")]
-    public virtual ICollection<Estudiante>? Estudiantes { get; set; }
-
-    [InverseProperty("Plantel")]
-    public virtual ICollection<Material>? Materiales { get; set; }
-}
