+++ InventoryControl/Program.Helpers.cs
