+++ InventoryControl/Program.Queries.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.ChangeTracking; // CollectionEntry
+using System;
+using static System.Console;
+using AlmacenDataContext;
+using AlmacenSQLiteEntities;
+
+partial class Program
+{
+
+}
