+++ InventoryControl/Program.cs
@@ -1,0 +1,15 @@
+﻿using AlmacenDataContext;
+using AlmacenSQLiteEntities;
+using System;
+using static System.Console;
+
+internal partial class Program
+{
+    static void Main()
+    {
+        Console.Clear();
+        Almacen db = new();
+        WriteLine($"Provider: {db.Database.ProviderName}");
+        UI.Manage();
+    }
+}
