--- InventoryControlApp.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33530.505
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InventoryControlApp", "InventoryControlApp.csproj", "{C6724257-6DA3-4361-AB1C-EABA7AC443AB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C6724257-6DA3-4361-AB1C-EABA7AC443AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C6724257-6DA3-4361-AB1C-EABA7AC443AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C6724257-6DA3-4361-AB1C-EABA7AC443AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C6724257-6DA3-4361-AB1C-EABA7AC443AB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {67E07303-E22B-436E-9C1F-2D2BCBDC509B}
-	EndGlobalSection
-EndGlobal
