+++ Libraries/AlmacenDataContext/AlmacenDataContext.cs
@@ -1,0 +1,6 @@
+﻿namespace AlmacenDataContext;
+
+public class AlmacenDataContext
+{
+
+}
