+++ Libraries/AlmacenSQLiteEntities/AlmacenSQLiteEntities.cs
@@ -1,0 +1,6 @@
+﻿namespace AlmacenSQLiteEntities;
+
+public class AlmacenSQLiteEntities
+{
+
+}
