+++ Libraries/AlmacenSQLiteEntities/BaseAlmacen.sql
