--- Program.Queries.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.ChangeTracking; // CollectionEntry
-using System;
-using static System.Console;
-
-partial class Program
-{
-
-}
